pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nemospawn"
version = "1.0.0"
description = "GPU-native agent-swarm orchestration for the full NVIDIA AI stack"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9",
    "rich>=13.0",
    "pynvml>=11.5",
    "psutil>=5.9",
    "pyyaml>=6.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
nemospawn = "nemospawn.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]
hpo = [
    "optuna>=4.0",
    "optuna-dashboard>=0.15",
]
transport = [
    "pyzmq>=25.0",
]
all = [
    "nemospawn[hpo,transport]",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nemospawn"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
